telegram-bot/requirements.txt
telethon==1.38.1
faust-streaming==0.11.3
openai==1.58.1